requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.7
python-dotenv>=1.0
httpx>=0.27

# Database
sqlalchemy>=2.0
alembic>=1.13

# Authentication
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7

# Payments
stripe>=10.0

# Validation
email-validator>=2.0
